created a separate component for badge, added badge to calendar, added hover and focus for calendar
