+++ src/components/CalendarPage/ChoosedDay/DayCalendarHead/DayCalendarHead.styled.jsx
@@ -1,6 +1,5 @@
 import styled from 'styled-components';
 import { backgroundColor, colorText } from 'src/styles/variables/Variables';
-import { Check } from 'src/components/shared/Icons';
 
 export const DayCalendarHeadStyles = styled.ul`
   padding: 14px 18px;
@@ -78,34 +77,4 @@ export const DateDay = styled.p`
   @media screen and (width >= 768px) {
     font-size: 16px;
   }
-`;
-
-export const TasksCount = styled.span`
-  position: absolute;
-  top: -7px;
-  left: -7px;
-
-  width: 15px;
-  height: 15px;
-  border-radius: 50%;
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  font-size: 12px;
-
-  color: white;
-  background: #ea3d65;
-`;
-
-export const CheckStyles = styled(Check)`
-  position: absolute;
-  top: -7px;
-  left: -7px;
-
-  width: 15px;
-  height: 15px;
-
-  transform: scale(1.25);
 `;
